factor out QueryCacheStatus.svelte
